core: do not append newlines to ouputs unless list/tuple (#34)
Instead of treating all stdout and stderr outputs as string lines
to have newlines, respect the stdout/stderr parameters provided to
register_subprocess().  This allows for more control of the fixture
and allow for binary inputs (and text data with line breaks other
than os.linesep).

Preserve the prior functionality when outputs are a list/tuple. In
this context, it makes sense to treat the inputs as a sequence of
text lines.

Update documentation to show both methods, add test cases for binary
data, and update existing tests that relied on the newline behavior.

Fixes #33

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>