subprocess: preserve command when returning FakePopen object (#37)
Use the calling command to populate the FakePopen object,
rather than the Command object.  This will prevent the Commmand
object from being visible in application code that does not
expect it, e.g. proc.args and CalledProcessError.cmd.

This is particularly noticeable on the application-side if using
fake_process.any() for matching args.

While here, update the type info for FakePopen.args (command)
property to match that found for Popen.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>